refactor main just a bit
